Add Startup Preferences (More info in Description)
Adds a prompt for users to set their beginning preferances. This only shows on the first start up (Right now it shows up all of the time because there is no way to write to the Config yet) It doesn't change any settings yet, because I need to be able to write to the config. Once there is a way to write, I will change the methods, and it will be fully functional. This also adds the 'prompt' node_module, which allows the server to prompt for user input. If you see any issues, let me know, or revert this commit.